more untangling.
