+++ pyserini/encode/__init__.py
@@ -20,16 +20,17 @@ from ._base import JsonlCollectionIterator, JsonlRepresentationWriter
 # from ._base import DocumentEncoder, QueryEncoder, JsonlCollectionIterator,\
 #     RepresentationWriter, FaissRepresentationWriter, JsonlRepresentationWriter, PcaEncoder
 from ._ance import AnceEncoder, AnceDocumentEncoder, AnceQueryEncoder
-# from ._auto import AutoQueryEncoder, AutoDocumentEncoder
+from ._auto import AutoQueryEncoder, AutoDocumentEncoder
 from ._dpr import DprDocumentEncoder, DprQueryEncoder
 from ._tct_colbert import TctColBertDocumentEncoder, TctColBertQueryEncoder
-# from ._aggretriever import AggretrieverDocumentEncoder, AggretrieverQueryEncoder
+from ._aggretriever import AggretrieverDocumentEncoder, AggretrieverQueryEncoder
 from ._unicoil import UniCoilEncoder, UniCoilDocumentEncoder, UniCoilQueryEncoder
 from ._cached_data import CachedDataQueryEncoder
 from ._tok_freq import TokFreqQueryEncoder
-# from ._splade import SpladeQueryEncoder
-# from ._slim import SlimQueryEncoder
-# from ._openai import OpenAIDocumentEncoder, OpenAIQueryEncoder, OPENAI_API_RETRY_DELAY
+from ._splade import SpladeQueryEncoder
+from ._slim import SlimQueryEncoder
 from ._cosdpr import CosDprEncoder, CosDprDocumentEncoder, CosDprQueryEncoder
+
+# from ._openai import OpenAIDocumentEncoder, OpenAIQueryEncoder, OPENAI_API_RETRY_DELAY
 # from ._clip import ClipEncoder, ClipDocumentEncoder
