+++ pyserini/encode/__main__.py
@@ -17,11 +17,12 @@
 import argparse
 import sys
 
-from pyserini.encode import JsonlRepresentationWriter, FaissRepresentationWriter, JsonlCollectionIterator
-from pyserini.encode import DprDocumentEncoder, TctColBertDocumentEncoder, AnceDocumentEncoder, AggretrieverDocumentEncoder, AutoDocumentEncoder, CosDprDocumentEncoder, ClipDocumentEncoder
+from pyserini.encode import JsonlRepresentationWriter, JsonlCollectionIterator
+from pyserini.encode._faiss import FaissRepresentationWriter
+from pyserini.encode import DprDocumentEncoder, TctColBertDocumentEncoder, AnceDocumentEncoder, AggretrieverDocumentEncoder, AutoDocumentEncoder, CosDprDocumentEncoder
 from pyserini.encode import UniCoilDocumentEncoder
-from pyserini.encode import OpenAIDocumentEncoder, OPENAI_API_RETRY_DELAY
-
+from pyserini.encode._openai import OpenAIDocumentEncoder, OPENAI_API_RETRY_DELAY
+from pyserini.encode._clip import ClipDocumentEncoder
 
 encoder_class_map = {
     "dpr": DprDocumentEncoder,
