+++ pyserini/search/hybrid/__init__.py
@@ -15,5 +15,3 @@
 #
 
 from ._searcher import HybridSearcher
-
-__all__ = ['HybridSearcher']
