+++ pyserini/search/nmslib/_searcher.py
@@ -13,6 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+
 import json
 import os
 from dataclasses import dataclass
